% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gconcord.R
\name{cv.gconcord}
\alias{cv.gconcord}
\title{Cross Validation for the Tuning Parameters in Graphical CONCORD Methods}
\usage{
cv.gconcord(
  data,
  rand,
  FUN,
  lam1.vec,
  lam2.vec,
  K = 3,
  method = "coordinatewise",
  tol = 1e-05,
  maxit = 100,
  steptype = 0
)
}
\arguments{
\item{data}{a numerical matrix or a data frame with n observations (rows) and p variables (columns).}

\item{rand}{optional, an integer vector of the length the number of observations (rows) in dataset,
assiging observations to different groups for cross validation. The default is that the function would
automatically create the integer vector under the argument \code{K}.}

\item{FUN}{optional, a user-defined cost function used in cross validation. It should contain two arguments:
\itemize{
       \item \code{param}: the estimated parameter, a matrix input of the inverse covariance matrix estimation.
       \item \code{data}: the data used in the cost function.
       }
       If \code{FUN} is messing, predictive risk function (see \code{\link{pred.risk}} for reference) will be used.
       That is, \code{FUN = pred.risk} by default.}

\item{lam1.vec}{optional, a numerical vector containing candidate values for L1 penalty parameter, or a p-by-p penalty matrix. In the
latter case, no cross validation will be conducted for L1 penalty parameter.}

\item{lam2.vec}{optional, a numerical vector containing candidate values for Frobenius norm penalty parameter.}

\item{K}{optional, the number of folds in cross validation. Default value is \code{3}.}

\item{method}{optional, a character string indicating which optimization method is used. See \code{\link{gconcord}} for reference.}

\item{tol}{optional, a numeric number for convergence threshold in optimization method. The default is \code{1e-05}. See \code{\link{gconcord}}.}

\item{maxit}{optional, an integer for the maximum number of iterations before termination in each optimization method.
The default is 100. See \code{\link{gconcord}}.}

\item{steptype}{optional, an integer determining the step type in \code{ista} and \code{fista}. See \code{\link{gconcord}}.}
}
\value{
Function \code{cv.gconcord} returns a list object containing 4 variables:
\itemize{
  \item{\code{val.error}}{: a validation error matrix for each
       candidate choice of lambda_1 and lambda_2. The row names and columns names are values of candidate choice of lambda_1
       and lambda_2, respectively. If lambda_1 is a penalty matrix, then the row name is denoted as -1.}
  \item{\code{val.error.quantile}}{: the quantile values in matrix for a validation error matrix \code{val.error}.}
  \item{\code{lam1.optimal}}{: optimal value for lambda_1.}
  \item{\code{lam2.optimal}}{: optimal value for lambda_2.}
}
}
\description{
For a given cost function, it estimates the optimal choice of the tuning parameters
for L1 penalty and Frobenius norm penalty in the Graphical CONCORD
optimization problem. This function can be used only if data are available.
}
\examples{
p = 10
n = 50
sigma = matrix(1,nrow = p,ncol = p)*0.5 + diag(1,p,p)*0.5
data <- round(mvrnorm(n, mu = rep(0,p), sigma),2)         # generate data

res <- cv.gconcord(data = data, lam1.vec = seq(0,1,0.02), # user-defined choices for lambda_1
                   lam2.vec = seq(0, 1, 0.02),            # user-defined choices for lambda_2
                   K = 5,                                 # 5-fold cross validation
                   method = "ista" )
res$lam1.optimal           # optimal lambda_1
res$lam2.optimal           # optimal lambda_2

}
