% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gconcord.R
\name{gconcord}
\alias{gconcord}
\title{Graphical CONvex CORrelation selection methoD (gconcord)}
\usage{
gconcord(
  data,
  S,
  lambda1,
  lambda2 = 0,
  method = "coordinatewise",
  tol = 1e-05,
  maxit = 100,
  steptype = 0
)
}
\arguments{
\item{data}{a numerical matrix or a data frame with n observations (rows) and
p variables (columns). \code{NULL} is set only if data are not available and cov will be used.}

\item{S}{optional, a symmetric numeric matrix. Sample covariance matrix is only required
if data are not available.}

\item{lambda1}{a non-negative scalar L-1 penalty parameter, or a p-by-p symmetric penalty matrix.}

\item{lambda2}{a non-negative scalar for L-2 penalty parameter. The default is 0.}

\item{method}{a character string indicating which optimization method is to be used.
It can be one of "\code{coordinatewise}", "\code{ista}", or "\code{fista}". The default is
"\code{coordinatewise}".}

\item{tol}{a numeric number for convergence threshold in optimization method. The default is \code{1e-05}.}

\item{maxit}{an integer for the maximum number of iterations before termination in each optimization method.
The default is 100.}

\item{steptype}{an integer determining the step type for \code{ista} and \code{fista}. Three available choices are:
\itemize{
\item \code{0}: initial step size is heuristic of Barzilai-Borwein. Please refere to \url{https://arxiv.org/pdf/1409.3768.pdf};
\item \code{1}: initial step size is 1;
\item \code{2}: initial step size is the feasible step size found from previous iteration.
}
The default value is \code{0}.}
}
\value{
A list object containing three variables:
\itemize{
\item\code{omega}: the computed optimal sparse inverse covariance estimator;
\item\code{nitr}: the total number of iterations in the computation;
\item\code{maxdiff}: a numeric vector presenting the covergence measures for all iterations.
}
}
\description{
Estimates a sparse inverse covariance matrix from a convex
pseudo-likelihood function with lasso L1 penalty or an elastic net penalty.
}
\details{
Coordinate-wise descent algorithm for gconcord is discussed in https://arxiv.org/pdf/1307.5381.pdf, both ISTA
and FISTA are discussed in https://arxiv.org/abs/1409.3768.
}
\examples{
# generate random data
p = 4
n = 100
sigma = matrix(1,nrow = p,ncol = p) * 0.5 + diag(1,p,p) * 0.5
set.seed(1)
library(MASS);data <- round(mvrnorm(n, mu = rep(0,p), sigma),2)

# scalar penalty parameters
omega <- gconcord(data, lambda1 = 0.36, lambda2 = 0.1,
                  method = "coordinatewise")
omega$omega    # the estimated sparse inverse covariance matrix
omega$nitr     # the number of iterations in the algorithm
omega$maxdiff  # the vector of the convergence measures


}
