% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gconcord.R
\name{get.data}
\alias{get.data}
\title{Get daily returns of DJIA component stocks for a given time horizon.}
\usage{
get.data(
  start = "1990-01-03",
  end = "2018-02-28",
  type = "return",
  na.rm = TRUE
)
}
\arguments{
\item{start}{optional, a string of starting date with format yyyy-mm-dd. The default
is "1990-01-03".}

\item{end}{optional, a string of ending date with format yyyy-mm-dd. The default is
"2018-02-28". Ending date should be no earlier than starting date.}

\item{type}{optional, a string indicating the type of DJIA data to be extracted. Defalut
value is \code{return}, the daily return of 30 component stocks of DJIA. Available choices are:
\itemize{
  \item{\code{return}}{: daily arithmatic return of all DJIA component stocks. Maximum time
  horizon is from 1990-01-03 to 2018-02-28.}
  \item{\code{price}}{: daily closing prices of all DJIA component stocks. Maximum time
  horizon is from 1990-01-02 to 2018-02-28.}
  \item{\code{index}}{: daily closing price of DJIA Index. Maximum time horizon is from
  1990-01-02 to 2018-03-02.}
}}

\item{na.rm}{optional, a boolean indicating if the column containing a missing value
is removed. The default is \code{TRUE}.}
}
\description{
Return a data frame containing daily returns of DJIA component stocks for a given
time horizon.
}
\examples{
# Get return data
data = get.data(start = "2017-01-01", end = "2017-12-31", type = "return", na.rm = TRUE)

# Get price data
data = get.data(start = "2005-04-15", end = "2005-04-30", type = "price", na.rm = FALSE)

# Get DJIA index data
data = get.data(start = "1998-10-27", end = "1999-01-01", type = "index")

}
