% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gconcord.R
\name{pred.risk}
\alias{pred.risk}
\title{Predictive Risk Loss Function in Cross Validation}
\usage{
pred.risk(omega, data)
}
\arguments{
\item{omega}{the estimated sparse inverse covariance matrix.}

\item{data}{data used in the loss function.}
}
\description{
Calculate the predictive risk for a given estimator \emph{omega}.
}
\examples{
p = 10
n = 50
sigma = matrix(1,nrow = p,ncol = p)*0.5 + diag(1,p,p)*0.5
data <- round(mvrnorm(n, mu = rep(0,p), sigma),2)         # generate data
omega <- gconcord(data, lambda1 = 0.2, lambda2 = 0.1, method = "ista")$omega
pred.risk(omega, data)

}
