% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gconcord.R
\name{zero.count}
\alias{zero.count}
\title{Get sparsity level of a matrix}
\usage{
zero.count(mat, ratio = TRUE, thres = 1e-05)
}
\arguments{
\item{mat}{a matrix input}

\item{ratio}{optional, a boolean value indicating if function returns a ratio of
number of zero elements over total number of elements, or the count of zero elements.
The default is \code{TRUE}.}

\item{thres}{optional, a threshold of determining a zero element. The default is 1e-5.}
}
\description{
Get the sparsity level of a matrix, measured by either a ratio of number of zero
elements over total number of elements, or the count of zero elements.
}
\examples{
mat <- diag(5)
zero.count(mat)
zero.count(mat, ratio = FALSE)

}
