% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DJIndex}
\alias{DJIndex}
\title{Daily DJIA index price from 1990/01/02 to 2018/02/28}
\format{A data frame with 7096 daily index price.}
\source{
https://finance.yahoo.com/
}
\usage{
DJIndex
}
\description{
A series data frame containing daily DJIA index
with 7096 observations.
}
\keyword{datasets}
