% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{return}
\alias{return}
\title{Daily arithmatic returns of 30 component stocks in DJIA index
from 1990/01/03 to 2018/02/28}
\format{A data frame with 7095 rows and 30 columns.
\describe{Airthmetic returns of stocks
    AAPL
  , AXP
  ,  BA
  , CAT
  , CSCO
  , CVX
  , DIS
  , DWDP
  ,  GE
  ,  GS
  ,  HD
  , IBM
  , INTC
  , JNJ
  , JPM
  ,  KO
  , MCD
  , MMM
  , MRK
  , MSFT
  , NKE
  , PFE
  ,  PG
  , TRV
  , UNH
  , UTX
  ,  V
  ,  VZ
  , WTM
  , XOM
}}
\source{
https://finance.yahoo.com/
}
\usage{
return
}
\description{
A 30-dimensional time series data frame containing
daily returns of 30 component stock in DJIA index
with 7095 observations and 30 columns.
}
\keyword{datasets}
